pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "iron-dillo"
version = "0.1.0"
description = "Watsonx-powered cybersecurity assistant for underserved communities"
readme = "README.md"
authors = [{name = "Iron Dillo Team"}]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=1.10,<3.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "umap-learn>=0.5",
    "numpy>=1.24",
    "python-multipart>=0.0.6",
]

[project.scripts]
iron-dillo = "iron_dillo.cli:main"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
